# openapi/schema.yaml
openapi: 3.0.3
info:
  title: Amazon Ads Optimizer API
  version: 0.1.0
  description: V1 account authentication, tenant/store/profile context and permission
    APIs.
paths:
...
  /api/v1/analysis/tasks:
    get:
      operationId: analysis_task_list
      tags:
      - analysis
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 分析任务列表
    post:
      operationId: api_v1_analysis_tasks_create
      tags:
      - analysis
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisCreateRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/AnalysisCreateRequest'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/AnalysisCreateRequest'
        required: true
      security:
      - bearerAuth: []
      responses:
        '202':
          description: 分析任务
  /api/v1/analysis/tasks/{task_id}:
    get:
      operationId: analysis_task_retrieve
      parameters:
      - in: path
        name: task_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - analysis
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 分析结果
...
  /api/v1/reports/tasks/{task_id}:
    get:
      operationId: api_v1_reports_tasks_retrieve
      summary: 查看导入任务、批次计数和行级错误
      parameters:
      - in: path
        name: task_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - reports
      security:
      - bearerAuth: []
      responses:
        '200':
          description: 导入任务详情
...
components:
  schemas:
...
    AnalysisCreateRequest:
      type: object
      properties:
        tenant_id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
      required:
      - profile_id
      - tenant_id
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT